Web modifications
There are numerous changes that I made to the website. From adding more web pages such as the landing page, user login/registration, a quiz page and a results page that will display the users results along with a congratulations text for completing the quiz.

I also implemented backend functionality, using Firebase as my primary database for this project. It consists of email/password authentication for the login/registration feature of the website, a results table with user ID, and a questions table. Initialy each questions were developed using typescript and was not stored in the database. For this project, it is very convinient to have all questions stored in the backend, so I had to figure out a way to do so. This was my first interaction with typescript, so I wasnt entirely sure as to how Im going to go about storing all questions in my backend. Thought of taking the Firestore Cloud Function method but given the time I had for this project, it didnt seem as though I was going to complete the work well on time.

My solution for this was to turn my typescript questions into a JSON file and import all questions into Firebase. I have imported texts, queries before but from Javascript file, not JSON. I had difficulties importing all questions to my Firebase and as of now, the questions that have been imported to my Firebase are visible on my database but not my website.